+++ modules/utils.py
@@ -17,19 +17,77 @@ limitations under the License.
 """
 
 # Imports
+from modules import ui, encryption
+
 import os
 import sys
+import uuid
+import string
+import random
+
+# Constants
+APP_VERSION = "v1.0.0"
 
 # Function 1: Set Terminal Title
-def set_terminal_title(app_version: str) -> None:
-    """
-    Sets the title of the terminal window.
-
-    Args:
-        app_version (String): The version of the app in SemVer.
-    """
+def set_terminal_title() -> None:
+    """ Sets the title of the terminal window. """
 
     if sys.platform.startswith("win"):
-        os.system(f"title Ghost Protocol {app_version}")
+        os.system(f"title Ghost Protocol {APP_VERSION}")
     else:
-        print(f"\033]0;Ghost Protocol {app_version}\a", end="", flush=True)
+        print(f"\033]0;Ghost Protocol {APP_VERSION}\a", end="", flush=True)
+
+# Function 2: Initiate Onboarding
+def initiate_onboarding(session_key_ready):
+    """
+    Initiates the onboarding process for the member.
+
+    Args:
+        session_key_ready (asyncio.Event()): Asynchronous signaling flag whether the session key is ready.
+    """
+
+    # Variables
+    session_key = None
+    room_id = None
+
+    # Displaying Initial Messages
+    ui.display_message(message="=== GHOST PROTOCOL (SECURE COMMS TERMINAL) ===", color="green", bright=True)
+    ui.display_message(message=f"Status: CLEAR // Protocol: Ghost-E2EE // Version: {APP_VERSION}", color="white", dim=True)
+    ui.display_message(message="[!] WARNING: Provided 'as is' without warranty. Use at your own risk.", color="red")
+
+    # Prompting User for Username & Room ID
+    username = ui.display_prompt(message="USERNAME: ", color="cyan", prefix="\n")
+
+    if not username:
+        ui.display_message(message="[!] ACCESS DENIED: Username cannot be empty.", color="red", prefix="\n")
+        return None
+
+    room_decision = ui.display_prompt(message="CREATE OR JOIN ROOM: ", color="cyan").lower()
+
+    if not room_decision or room_decision not in ["create", "join"]:
+        ui.display_message(message="[!] ACCESS DENIED: Enter a valid input ('create' or 'join').", color="red", prefix="\n")
+        return None
+
+    if room_decision == "create":
+        room_id = "".join(random.choices(string.ascii_lowercase + string.digits, k=8))
+        ui.display_message(message=f"ROOM ID: {room_id}", color="white", dim=True, prefix="\n")
+
+        session_key = encryption.generate_key(length=256)
+        session_key_ready.set()
+    elif room_decision == "join":
+        room_id = ui.display_prompt(message="ROOM ID: ", color="cyan")
+        print()
+
+    if not room_id or len(room_id) != 8:
+        ui.display_message(message="[!] ACCESS DENIED: Room ID is required and must be exactly 8 characters.", color="red")
+        return None
+
+    # Generating the Client ID (UUID v4)
+    client_id = str(uuid.uuid4())
+
+    # Displaying Client Information
+    ui.display_message(message=f"SHORT CLIENT ID: {client_id.split('-')[4]}", color="white", dim=True)
+    ui.display_message(message=f"CLIENT ID: {client_id}", color="white", dim=True)
+    ui.display_message(message="[*] Requesting authentication token and establishing connection...", color="yellow", prefix="\n")
+
+    return username.replace(" ", "-"), room_decision, room_id, session_key, client_id